Create README.md
Click lower left of screen and 3 circles of position are drawn in red, green, blue with the stars, Arcturus, Vega, Dubhe at the centers.  Click the red square in the lower right, labeled clock, and now you are inside the sphere.  Move the mouse to orient the 3 COP's so that only one set is visible.  As time progresses, one set of COP's remains fixed and the other set emerges as time progresses.  Click the upper left of the screen to only see the moveable set of COP's.  The white circle represents the latitude at which the COP's are plausible.  Notice the intersection of the COP's remains on the circle of latitude.  It is the longitude of the COP intersection that varies.